Need a distributed `solve_for_pressure`
