+++ sandbox/mpi_turbulence.jl
@@ -1,5 +1,3 @@
-include("distributed_model.jl")
-
 using MPI
 
 MPI.Initialized() || MPI.Init()
@@ -13,33 +11,6 @@ using Oceananigans.AbstractOperations
 using Oceananigans.Utils
 
 using Oceananigans.Solvers: calculate_pressure_right_hand_side!, copy_pressure!
-
-import Oceananigans.Solvers: solve_for_pressure!
-
-child_architecture(::CPU) = CPU()
-
-function solve_for_pressure!(pressure, solver::DistributedFFTBasedPoissonSolver, arch, grid, Δt, U★)
-
-    RHS = first(solver.storage)
-
-    rhs_event = launch!(arch, grid, :xyz,
-                        calculate_pressure_right_hand_side!, RHS, arch, grid, Δt, U★,
-                        dependencies = Event(device(arch)))
-
-    wait(device(arch), rhs_event)
-
-    solve_poisson_equation!(solver)
-
-    ϕ = first(solver.storage)
-
-    copy_event = launch!(arch, grid, :xyz,
-                         copy_pressure!, pressure, ϕ, arch, grid,
-                         dependencies = Event(device(arch)))
-
-    wait(device(arch), copy_event)
-
-    return nothing
-end
 
 topo = (Periodic, Periodic, Periodic)
 full_grid = RegularRectilinearGrid(topology=topo, size=(512, 512, 1), extent=(4π, 4π, 1), halo=(3, 3, 3))
